requirements.txt
Flask~=1.1
flask-cors==3.0.10
Flask-JSONRPC~=1.1.0
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth==1.27.0
hdwallet~=0.2.0
pip~=21.0.1
py_crypto_hd_wallet
pyodbc~=4.0.30
PyMySQL==1.0.2
redis==3.2.1
simplejson==3.17.2
SQLAlchemy==1.4.3
web3~=5.15.0
#only need this for IIS development
#steps here https://pypi.org/project/wfastcgi/
#wfastcgi~=3.0.0
#if required for additional firebase features
#firebase-admin